Changed models and implemented functionality for adding ingredients/recipes

+++ back-end/YummyGen/YummyGen.Domain/Interfaces/IRepository.cs
@@ -5,7 +5,7 @@ namespace YummyGen.Domain.Interfaces
     public interface IRepository<T> where T : class
     {
         public Task<IEnumerable<T>> GetAll();
-        public Task<T> GetById(long id);
+        public Task<T> GetById(int id);
         public Task<T> Add(T entity);
         public Task<T> Update(T entity);
         public Task Delete(T entity);
